requirements.txt
-e .
aws_cdk.aws_ecs
aws_cdk.aws_certificatemanager
cdk_fargate_patterns